course-chatbot copy/requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
google-generativeai==0.3.0
python-dotenv==1.0.0
